+++ src/main/resources/templates/main/main.html
@@ -7,19 +7,20 @@
     <link rel="stylesheet" type="text/css" href="/css/styles.css">
     <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
     <script src="/js/menu.js"></script>
-    <script src="/js/user_search.js"></script>
+    <script type="text/javascript" th:src="@{/js/user_search.js}"></script>
     <title>Title</title>
     <link rel="icon" href="image/favicon.ico" type="image/x-icon">
 </head>
 <body>
-<div th:replace="fragments/header :: menu1">메뉴1부분</div>
+<div th:replace="~{fragments/header :: menu1}">메뉴1부분</div>
 
-<div th:replace="fragments/header :: menu2">메뉴2부분</div>
+<div th:replace="~{fragments/header :: menu2}">메뉴2부분</div>
 
 <div>
     <h1>테스트</h1>
 </div>
 
-<div th:replace="fragments/footer :: company">기업설명부분</div>
+<div th:replace="~{fragments/footer :: company}">기업설명부분</div>
+
 </body>
 </html>
